throw custom error types (#921)
* throw custom error types

We want to be able to show nice error messages to the user so this
introduces a CedarException type that we can catch in CedarEDA and print
without the stacktrace.

There are however two types of errors, we fked up errors and you fked up
error. When there is an error in the user's circuit we want to show a
friendly message, but when we hit an internal error I think the correct
behaviour is to show a stacktrace.

I took a rough guess at which errors we should wrap in a Cedar type and
which shouldn't. Some review may be in order.

* update unimplemented device test

* Suppress backtrace for CedarExceptions

* Fix module name in test

* type-parameter WrappedCedarException on what is wrapped and update tests

---------

Co-authored-by: Frames White <me@oxinabox.net>